changed RadioButton color, and renamed to just RadioButton
