+++ app/views/AutoStartPositionView.tsx
@@ -1,7 +1,7 @@
 import { Text, View, StyleSheet } from 'react-native';
 import { useState } from 'react';
 import { Switch } from 'react-native-paper';
-import RadioButtonComponent from '../components/RadioButtonComponent';
+import RadioButtonComponent from '../components/RadioButton';
 
 function AutoStartPositionView () {
     const [isEnabled, setIsEnabled] = useState(false);
